+++ scribejava-apis/src/main/java/com/github/scribejava/apis/ImgurApi.java
@@ -6,6 +6,8 @@ import com.github.scribejava.core.httpclient.HttpClient;
 import com.github.scribejava.core.httpclient.HttpClientConfig;
 import com.github.scribejava.core.model.OAuthConstants;
 import com.github.scribejava.core.model.ParameterList;
+
+import java.io.OutputStream;
 import java.util.Map;
 
 public class ImgurApi extends DefaultApi20 {
@@ -55,9 +57,10 @@ public class ImgurApi extends DefaultApi20 {
 
     @Override
     public ImgurOAuthService createService(String apiKey, String apiSecret, String callback, String defaultScope,
-            String responseType, String userAgent, HttpClientConfig httpClientConfig, HttpClient httpClient) {
-        return new ImgurOAuthService(this, apiKey, apiSecret, callback, defaultScope, responseType, userAgent,
-                httpClientConfig, httpClient);
+            String responseType, OutputStream debugStream, String userAgent, HttpClientConfig httpClientConfig,
+            HttpClient httpClient) {
+        return new ImgurOAuthService(this, apiKey, apiSecret, callback, defaultScope, responseType, debugStream,
+                userAgent, httpClientConfig, httpClient);
     }
 
     public static boolean isOob(String callback) {
