+++ scribejava-apis/src/main/java/com/github/scribejava/apis/wunderlist/WunderlistOAuthService.java
@@ -1,5 +1,7 @@
 package com.github.scribejava.apis.wunderlist;
 
+import java.io.OutputStream;
+
 import com.github.scribejava.apis.WunderlistAPI;
 import com.github.scribejava.core.httpclient.HttpClient;
 import com.github.scribejava.core.httpclient.HttpClientConfig;
@@ -9,9 +11,10 @@ import com.github.scribejava.core.oauth.OAuth20Service;
 public class WunderlistOAuthService extends OAuth20Service {
 
     public WunderlistOAuthService(WunderlistAPI api, String apiKey, String apiSecret, String callback,
-            String defaultScope, String responseType, String userAgent, HttpClientConfig httpClientConfig,
-            HttpClient httpClient) {
-        super(api, apiKey, apiSecret, callback, defaultScope, responseType, userAgent, httpClientConfig, httpClient);
+            String defaultScope, String responseType, OutputStream debugStream, String userAgent,
+            HttpClientConfig httpClientConfig, HttpClient httpClient) {
+        super(api, apiKey, apiSecret, callback, defaultScope, responseType, debugStream, userAgent, httpClientConfig,
+                httpClient);
     }
 
     @Override
