+++ scribejava-core/src/test/java/com/github/scribejava/core/AbstractClientTest.java
@@ -44,7 +44,8 @@ public abstract class AbstractClientTest {
 
     @Before
     public void setUp() {
-        oAuthService = new OAuth20Service(null, "test", "test", null, null, null, null, null, createNewClient());
+        oAuthService = new OAuth20Service(null, "test", "test", null, null, null, System.out, null, null,
+                createNewClient());
     }
 
     @After
